[#26] modify: Header 및 Footer container 의 위치 이동
Title >
* Header 및 Footer container 의 위치 이동

Description >
* AppRouter 에서 렌더링 되던 Header 및 Footer Container 의 위치를
  App 으로 변경.
  (AppRouter 에서 Header, Footer 를 렌더링하는것은 적절하지
  않다고 판단)
