+++ client/src/routers/AppRouter.jsx
@@ -12,30 +12,9 @@ import MilestoneCreatePage from '@Pages/MilestoneCreatePage';
 import MilestoneListPage from '@Pages/MilestoneListPage';
 import MilestoneModifyPage from '@Pages/MilestoneModifyPage';
 import LoginPage from '@Pages/LoginPage';
-import Cookies from 'universal-cookie';
-
-import { setLoginInfo, resetLoginInfo } from '../modules/login';
-import { useDispatch } from 'react-redux';
-
-import HeaderContainer from '@Containers/HeaderContainer'
-import FooterContainer from '@Containers/FooterContainer'
 
 function AppRouter() {
-  const dispatch = useDispatch();
-
-  const cookies = new Cookies();
-  const loginInfo = cookies.get('jwt');
-  
-  if (loginInfo !== undefined) {
-    dispatch(setLoginInfo(loginInfo));
-  }
-  else {
-    dispatch(resetLoginInfo());
-  }
-
   return (
-    <>
-    <HeaderContainer />
     <Router>
       <Route exact={true} path="/" component={IssueListPage} />
       <Route path="/issue-detail/:issueId/" component={IssueDetailPage} />
@@ -46,8 +25,6 @@ function AppRouter() {
       <Route path="/milestone-modify/:milestoneId/" component={MilestoneModifyPage} />
       <Route path="/login/" component={LoginPage} />
     </Router>
-    <FooterContainer />
-    </>
   );
 }
